+++ Makcha/src/pages/WaitingSpot/common/CategoryTab.tsx
