+++ Makcha/src/pages/WaitingSpot/common/FooterButton.tsx
