+++ Makcha/src/pages/WaitingSpot/common/WaitingSpotHeader.tsx
