+++ Makcha/src/pages/WaitingSpot/common/constants.ts
@@ -11,3 +11,9 @@ export const waitingCategories = [
   };
 
 export const FALLBACK_CENTER = { lat: 37.5665, lng: 126.9780 }; //임시 좌표
+
+export const routeCategories = [
+    { key: "shortest", label: "최단" },
+    { key: "safe", label: "안전우선" },
+    { key: "bright", label: "밝은길" },
+  ] as const;
