+++ Makcha/src/pages/WaitingSpot/components/DirectionList.tsx
@@ -1,4 +1,4 @@
-import type { DirectionListProps } from "../../types/walking-direction";
+import type { DirectionListProps } from "../../../types/walking-direction";
 import { DirectionCard } from "./DirectionCard"
 
 export const DirectionList = ({direction} : DirectionListProps) => {
